data-backend/requirement.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-use-email-as-username==1.2.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
Markdown==3.4.1
mysqlclient==2.1.1
pipreqs==0.4.11
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
yarg==0.1.9
